+++ source/extensions/omni.flux.material_api/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.1.1]
+## Changed
+- tweaked "IsHidden" behavior to only count "hidden" metadata and not "unused"
+
 ## [0.1.0]
 ### Added
 - Created. Extracted python api from omni.kit.property.materials-1.10.17 + omni.kit.property.usd-4.2.16.